requirements.txt
# ChaosBench-Logic Dependencies
# Install with: pip install -r requirements.txt

# LLM API Clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
huggingface-hub>=0.20.0

# Data Processing & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
